# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelX API",
    "description": "\nSecurity operations platform: ingestion, detection, correlation, investigation.\n\n**Authentication.** `POST /api/auth/login` returns a short-lived access token; send it as\n`Authorization: Bearer <token>`. A refresh token is set as an httpOnly cookie and exchanged at\n`POST /api/auth/refresh`. Roles: VIEWER (read only), ANALYST (+ investigation work), ADMIN\n(+ users, rules, settings, audit log). The role each route needs is in docs/api.md.\n\n**Lists** return `{\"items\", \"total\", \"limit\", \"offset\"}`.\n\n**Errors** always have the shape\n`{\"error\": {\"code\", \"message\", \"request_id\", \"details\"?}}`. Validation errors (422) never echo\nsubmitted values.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dashboard/summary": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Summary",
        "description": "Records processed, events (total and today), alerts (today, open by severity), incidents\n(open, under investigation), monitored hosts, active rules, the top rules and source\naddresses of the last 7 days, and the latest alerts and incidents. Times are UTC.",
        "operationId": "get_summary_api_dashboard_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but the role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/dashboard/trends": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Trends",
        "description": "Per UTC day, oldest first, with every day present: alerts created (by severity),\nincidents created, and events (by their own time).",
        "operationId": "get_trends_api_dashboard_trends_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trends"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but the role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AlertStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "TRIAGED",
          "IN_PROGRESS",
          "RESOLVED",
          "FALSE_POSITIVE"
        ],
        "title": "AlertStatus"
      },
      "AlertSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          },
          "disposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Disposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority_score": {
            "type": "integer",
            "title": "Priority Score"
          },
          "priority_band": {
            "type": "string",
            "title": "Priority Band"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "destination_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Ip"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "evidence_truncated": {
            "type": "boolean",
            "title": "Evidence Truncated"
          },
          "detection_count": {
            "type": "integer",
            "title": "Detection Count"
          },
          "first_event_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Event At"
          },
          "last_event_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Event At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rule_id",
          "title",
          "severity",
          "confidence",
          "status",
          "disposition",
          "priority_score",
          "priority_band",
          "host",
          "username",
          "source_ip",
          "destination_ip",
          "event_count",
          "evidence_truncated",
          "detection_count",
          "first_event_at",
          "last_event_at",
          "created_at",
          "updated_at",
          "simulated"
        ],
        "title": "AlertSummary",
        "description": "One row of the alert queue."
      },
...
      "AlertsBySeverity": {
        "properties": {
          "critical": {
            "type": "integer",
            "title": "Critical"
          },
          "high": {
            "type": "integer",
            "title": "High"
          },
          "medium": {
            "type": "integer",
            "title": "Medium"
          },
          "low": {
            "type": "integer",
            "title": "Low"
          }
        },
        "type": "object",
        "required": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "AlertsBySeverity"
      },
...
      "DashboardSummary": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "records_processed": {
            "type": "integer",
            "title": "Records Processed"
          },
          "events_stored": {
            "type": "integer",
            "title": "Events Stored"
          },
          "events_today": {
            "type": "integer",
            "title": "Events Today"
          },
          "alerts_today": {
            "type": "integer",
            "title": "Alerts Today"
          },
          "open_alerts": {
            "type": "integer",
            "title": "Open Alerts"
          },
          "open_alerts_simulated": {
            "type": "integer",
            "title": "Open Alerts Simulated"
          },
          "critical_alerts": {
            "type": "integer",
            "title": "Critical Alerts"
          },
          "high_alerts": {
            "type": "integer",
            "title": "High Alerts"
          },
          "severity_distribution": {
            "items": {
              "$ref": "#/components/schemas/SeverityCount"
            },
            "type": "array",
            "title": "Severity Distribution"
          },
          "open_incidents": {
            "type": "integer",
            "title": "Open Incidents"
          },
          "incidents_under_investigation": {
            "type": "integer",
            "title": "Incidents Under Investigation"
          },
          "monitored_hosts": {
            "type": "integer",
            "title": "Monitored Hosts"
          },
          "inventory_assets": {
            "type": "integer",
            "title": "Inventory Assets"
          },
          "active_rules": {
            "type": "integer",
            "title": "Active Rules"
          },
          "library_rules": {
            "type": "integer",
            "title": "Library Rules"
          },
          "top_rules": {
            "items": {
              "$ref": "#/components/schemas/TopRule"
            },
            "type": "array",
            "title": "Top Rules"
          },
          "top_source_ips": {
            "items": {
              "$ref": "#/components/schemas/TopSource"
            },
            "type": "array",
            "title": "Top Source Ips"
          },
          "recent_alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertSummary"
            },
            "type": "array",
            "title": "Recent Alerts"
          },
          "recent_incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummary"
            },
            "type": "array",
            "title": "Recent Incidents"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "records_processed",
          "events_stored",
          "events_today",
          "alerts_today",
          "open_alerts",
          "open_alerts_simulated",
          "critical_alerts",
          "high_alerts",
          "severity_distribution",
          "open_incidents",
          "incidents_under_investigation",
          "monitored_hosts",
          "inventory_assets",
          "active_rules",
          "library_rules",
          "top_rules",
          "top_source_ips",
          "recent_alerts",
          "recent_incidents"
        ],
        "title": "DashboardSummary",
        "description": "Every number is counted in the database when requested (definitions in docs/api.md)."
      },
      "Disposition": {
        "type": "string",
        "enum": [
          "confirmed_malicious",
          "benign_expected"
        ],
        "title": "Disposition",
        "description": "Why a RESOLVED alert was closed (brief \u00a745: confirmed / resolved)."
      },
...
      "ErrorDetail": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ErrorDetail"
      },
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The shape of every error response (app/core/errors.py)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "TRIAGED",
          "INVESTIGATING",
          "CONTAINED",
          "RESOLVED",
          "CLOSED"
        ],
        "title": "IncidentStatus"
      },
      "IncidentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "risk_band": {
            "type": "string",
            "title": "Risk Band"
          },
          "alert_count": {
            "type": "integer",
            "title": "Alert Count"
          },
          "hosts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hosts"
          },
          "usernames": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Usernames"
          },
          "source_ips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ips"
          },
          "tactics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tactics"
          },
          "first_activity_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Activity At"
          },
          "last_activity_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "assigned_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "title",
          "status",
          "severity",
          "risk_score",
          "risk_band",
          "alert_count",
          "hosts",
          "usernames",
          "source_ips",
          "tactics",
          "first_activity_at",
          "last_activity_at",
          "created_at",
          "updated_at",
          "simulated"
        ],
        "title": "IncidentSummary"
      },
...
      "SeverityCount": {
        "properties": {
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "count"
        ],
        "title": "SeverityCount"
      },
...
      "TopRule": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "alerts": {
            "type": "integer",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "name",
          "alerts"
        ],
        "title": "TopRule"
      },
      "TopSource": {
        "properties": {
          "source_ip": {
            "type": "string",
            "title": "Source Ip"
          },
          "alerts": {
            "type": "integer",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "source_ip",
          "alerts"
        ],
        "title": "TopSource"
      },
...
      "TrendDay": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "alerts": {
            "$ref": "#/components/schemas/AlertsBySeverity"
          },
          "incidents": {
            "type": "integer",
            "title": "Incidents"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "day",
          "alerts",
          "incidents",
          "events"
        ],
        "title": "TrendDay"
      },
      "Trends": {
        "properties": {
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrendDay"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "days",
          "items"
        ],
        "title": "Trends"
      },
...
      "UserRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "UserRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}